+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Change log
 
+## Next
+> fix: use notifyListeners after ChangeNotifier disposed. Thanks laiiihz for [PR#555](https://github.com/xuelongqy/flutter_easy_refresh/pull/555).  
+> feat: ClassicHeader、ClassicFooter add IconThemeData. Thanks Lay523 for [PR#562](https://github.com/xuelongqy/flutter_easy_refresh/pull/562).  
+> feat: ClassicIndicator add [progressIndicatorSize] and [progressIndicatorStrokeWidth].  
+
 ## V 3.0.0+3
 > fix: dart >=2.13.0.  
 > fix: The screen is not full, [infinite] can not reset.  
-> feat: HeaderLocator and FooterLocator add clearExtent.  
+> feat: HeaderLocator and FooterLocator add [clearExtent].  
 > feat: Add OverrideFooter and OverrideHeader.  
 
 ## V 3.0.0+2